livekit-agent/requirements.txt
livekit-agents[google,silero,turn-detector]>=1.0
python-dotenv
aiohttp
